+++ back/src/main/java/com/ms/back/member/domain/port/RefreshService.java
@@ -1,0 +1,10 @@
+package com.ms.back.member.domain.port;
+
+import com.ms.back.member.domain.model.RefreshEntity;
+
+public interface RefreshService {
+    boolean existsByRefresh(String refresh);
+    void deleteByRefresh(String refresh);
+    void deleteByUsername(String username);
+    RefreshEntity save(RefreshEntity refreshEntity);
+}
