--- back/src/main/java/com/ms/back/member/persentation/controller/AdminController.java
@@ -1,16 +1,0 @@
-package com.ms.back.member.persentation.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-@Controller
-@ResponseBody
-public class AdminController {
-
-    @GetMapping("/admin")
-    public String adminP() {
-
-        return "admin Controller";
-    }
-}
