+++ back/src/main/java/com/ms/back/member/persentation/controller/AuthController.java
@@ -1,41 +1,164 @@
 package com.ms.back.member.persentation.controller;
 
+import com.ms.back.member.application.dto.JoinDTO;
 import com.ms.back.member.application.dto.LoginRequestDTO;
 import com.ms.back.member.application.dto.LoginResponseDTO;
+import com.ms.back.member.application.service.JoinService;
 import com.ms.back.member.application.service.LoginService;
+import com.ms.back.global.jwt.JWTUtil;
+import com.ms.back.member.domain.model.RefreshEntity;
+import com.ms.back.member.infrastructure.repository.RefreshRepository;
+import io.jsonwebtoken.ExpiredJwtException;
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Collection;
+import java.util.Date;
+import java.util.Iterator;
+
 @Controller
+@ResponseBody
 public class AuthController {
 
-    private final LoginService loginService;
+    private static final Logger logger = LoggerFactory.getLogger(AuthController.class);
 
-    public AuthController(LoginService loginService) {
+    private final LoginService loginService;
+    private final JoinService joinService;
+    private final JWTUtil jwtUtil;
+    private final RefreshRepository refreshRepository;
+
+    public AuthController(LoginService loginService,
+                          JoinService joinService,
+                          JWTUtil jwtUtil,
+                          RefreshRepository refreshRepository) {
         this.loginService = loginService;
+        this.joinService = joinService;
+        this.jwtUtil = jwtUtil;
+        this.refreshRepository = refreshRepository;
     }
 
-    // 로그인 페이지 보여주기
+    // 로그인 페이지 보여주기 (뷰 리턴)
     @GetMapping("/login")
     public String loginPage() {
-        return "login"; // 뷰 이름 (login.html 등)
+        return "login";
     }
 
     // 로그인 API
     @PostMapping("/login")
-    @ResponseBody
     public ResponseEntity<LoginResponseDTO> login(@RequestBody LoginRequestDTO request, HttpServletResponse response) {
         LoginResponseDTO loginResponse = loginService.login(request, response);
-
         if (loginResponse == null) {
             return ResponseEntity.status(401).build();
         }
-
         response.setHeader("access", loginResponse.getAccessToken());
-        // refresh token 쿠키 세팅 등은 loginService 내부에서 처리 가능
-
         return ResponseEntity.ok(loginResponse);
+    }
+
+    // 회원가입 API
+    @PostMapping("/join")
+    public ResponseEntity<String> joinProcess(@RequestBody JoinDTO joinDTO) {
+        logger.info("🟢 /join 요청 들어옴: {}", joinDTO);
+        boolean result = joinService.joinProcess(joinDTO);
+        if (result) {
+            logger.info("✅ 회원가입 성공: {}", joinDTO.getUsername());
+            return ResponseEntity.ok("회원가입 성공");
+        } else {
+            logger.warn("⚠ 이미 존재하는 아이디: {}", joinDTO.getUsername());
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("이미 존재하는 아이디입니다");
+        }
+    }
+
+    // 메인 페이지 API (인증 사용자 정보 확인용)
+    @GetMapping("/")
+    public String mainP() {
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Collection<? extends GrantedAuthority> authorities = authentication.getAuthorities();
+        Iterator<? extends GrantedAuthority> iter = authorities.iterator();
+        GrantedAuthority auth = iter.next();
+        String role = auth.getAuthority();
+
+        return "Main Controller: " + username + " / Role: " + role;
+    }
+
+    // 어드민 페이지 테스트용
+    @GetMapping("/admin")
+    public String adminP() {
+        return "admin Controller";
+    }
+
+    // 토큰 재발급 API
+    @PostMapping("/reissue")
+    public ResponseEntity<?> reissue(HttpServletRequest request, HttpServletResponse response) {
+        String refresh = null;
+        Cookie[] cookies = request.getCookies();
+        if (cookies != null) {
+            for (Cookie cookie : cookies) {
+                if ("refresh".equals(cookie.getName())) {
+                    refresh = cookie.getValue();
+                }
+            }
+        }
+
+        if (refresh == null) {
+            return new ResponseEntity<>("refresh token null", HttpStatus.BAD_REQUEST);
+        }
+
+        try {
+            jwtUtil.isExpired(refresh);
+        } catch (ExpiredJwtException e) {
+            return new ResponseEntity<>("refresh token expired", HttpStatus.BAD_REQUEST);
+        }
+
+        String category = jwtUtil.getCategory(refresh);
+        if (!"refresh".equals(category)) {
+            return new ResponseEntity<>("invalid refresh token", HttpStatus.BAD_REQUEST);
+        }
+
+        Boolean isExist = refreshRepository.existsByRefresh(refresh);
+        if (!isExist) {
+            return new ResponseEntity<>("invalid refresh token", HttpStatus.BAD_REQUEST);
+        }
+
+        String username = jwtUtil.getUsername(refresh);
+        String role = jwtUtil.getRole(refresh);
+
+        String newAccess = jwtUtil.createJwt("access", username, role, 600000L);
+        String newRefresh = jwtUtil.createJwt("refresh", username, role, 86400000L);
+
+        refreshRepository.deleteByRefresh(refresh);
+        addRefreshEntity(username, newRefresh, 86400000L);
+
+        response.setHeader("access", newAccess);
+        response.addCookie(createCookie("refresh", newRefresh));
+
+        return new ResponseEntity<>(HttpStatus.OK);
+    }
+
+    private Cookie createCookie(String key, String value) {
+        Cookie cookie = new Cookie(key, value);
+        cookie.setMaxAge(24*60*60);
+        cookie.setHttpOnly(true);
+        cookie.setPath("/");
+        return cookie;
+    }
+
+    private void addRefreshEntity(String username, String refresh, Long expiredMs) {
+        Date date = new Date(System.currentTimeMillis() + expiredMs);
+        RefreshEntity refreshEntity = new RefreshEntity();
+        refreshEntity.setUsername(username);
+        refreshEntity.setRefresh(refresh);
+        refreshEntity.setExpiration(date.toString());
+        refreshRepository.save(refreshEntity);
     }
 }
