--- back/src/main/java/com/ms/back/member/persentation/controller/MainController.java
@@ -1,31 +1,0 @@
-package com.ms.back.member.persentation.controller;
-
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-import java.util.Collection;
-import java.util.Iterator;
-
-@Controller
-@ResponseBody
-public class MainController {
-
-    @GetMapping("/")
-    public String mainP() {
-
-        String username = SecurityContextHolder.getContext().getAuthentication().getName();
-
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-
-        Collection<? extends GrantedAuthority> authorities = authentication.getAuthorities();
-        Iterator<? extends GrantedAuthority> iter = authorities.iterator();
-        GrantedAuthority auth = iter.next();
-        String role = auth.getAuthority();
-
-        return "Main Controller" + username + role;
-    }
-}
